+++ Data_validation.py
@@ -1,7 +1,8 @@
 import os
 # Need to install first time manually
 # os.system('install_packages.bat')
-import connections
+import pyodbc as odbc
+import snowflake.connector
 import pandas as pd
 import io
 import csv
@@ -14,6 +15,34 @@ from input_server_details import global_result
 # Suppress specific warnings
 warnings.simplefilter("ignore", UserWarning)
 warnings.simplefilter("ignore", FutureWarning)
+
+def setup_SQLSERVER_connection(server_name, login_id, password):
+
+    if login_id and password:
+        connection_string = (
+            f"Driver={{ODBC Driver 17 for SQL Server}};"
+            f"Server={server_name};"
+            f"UID={login_id};"
+            f"PWD={password};"
+        )
+    else:
+        connection_string = (
+            f"Driver={{ODBC Driver 17 for SQL Server}};"
+            f"Server={server_name};"
+            f"Trusted_Connection=Yes"
+        )
+    connection = odbc.connect(connection_string)
+    return connection
+
+def setup_SNOWFLAKE_connection(username, password, account, warehouse):
+    snowflake_config = {
+        "user": username,
+        "password": password,
+        "account": account,
+        "warehouse": warehouse
+    }
+    connection = snowflake.connector.connect(**snowflake_config)
+    return connection
 
 def get_datatype(connection, db, schema, table_name):
     column_datatypes = pd.read_sql(f" SELECT UPPER(COLUMN_NAME) as COLUMN_NAME, UPPER(DATA_TYPE) as DATA_TYPE"
